컨텐츠 프로그래밍 및 OIT 개선 작업.
컨텐츠 프로그래밍 및 OIT 개선 작업.

+++ DUOLClient/include/DUOLClient/ECS/Component/Enemy/EnemyGroupController.h
@@ -53,7 +53,7 @@ namespace DUOLClient
 
 		int _eliteEnemyCount;
 
-		int _createWaitForSeconds;
+		float _createWaitForSeconds;
 
 		DUOLMath::Vector3 _enemyGroupCenterPos;
